backend-python/requirements.txt
Django>=3.2 # 
djongo>=1.3.6 # 
djangorestframework>=3.12 # 3.12.4
django-cors-headers>=3.13 # 3.13.0

pymongo>=3.11 # 3.11.4
arrow>=1.3 # 1.3.0
termcolor>=2.5 # 2.5.0

pandas
scipy
aiohttp
tqdm
pyttsx3
dotenv
